requirements.txt
# Core Flask Framework
Flask>=3.1.1
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Flask-WTF>=1.2.2
WTForms>=3.2.1
Werkzeug>=3.1.3

# Database
SQLAlchemy>=2.0.42
psycopg2-binary>=2.9.10

# Authentication & Security
passlib[bcrypt]>=1.7.4
PyJWT>=2.10.1
python-jose[cryptography]>=3.5.0
jose>=1.0.0

# Image Processing
Pillow>=11.3.0

# AI Services
google-genai>=1.28.0

# Payment Processing
stripe>=12.4.0

# Email Validation
email-validator>=2.2.0

# OAuth (Optional)
Flask-Dance>=7.1.0
oauthlib>=3.3.1

# Development & Testing
gunicorn[gthread]>=23.0.0

# Additional Utilities
sift-stack-py>=0.8.2
python-multipart>=0.0.20
pydantic-settings>=2.10.1
pydantic>=2.11.7

# Template Engine
Jinja2>=3.1.6

# Development Server
uvicorn>=0.35.0
a2wsgi>=1.10.10 